renamed defaultContext to context to remain consistent with RequestTrigger function
